pyproject.toml
[project]
name = "tz-vanger"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-admin-sortable2>=2.3.1",
    "django-cors-headers>=4.9.0",
    "django-filer>=3.4.4",
    "django-mptt>=0.18.0",
    "mysqlclient>=2.2.8",
    "pre-commit>=4.5.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 119
exclude = [
    ".git",
    ".gitignore",
    "__pycache__",
    "__init__.py",
    "**/migrations",
]
target-version = "py312"
fix = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
ignore = ["W293", "A003"]
select = ["E", "F", "I", "B", "C", "A", "N", "UP"]
mccabe.max-complexity = 12
